+++ include/xqilla/ast/XQPredicate.hpp
@@ -57,7 +57,7 @@ protected:
 class XQILLA_API PredicateFilterResult : public ResultImpl
 {
 public:
-	PredicateFilterResult(const Result &parent, const ASTNode *pred, unsigned int contextSize);
+	PredicateFilterResult(const Result &parent, const ASTNode *pred, size_t contextSize);
 	Item::Ptr next(DynamicContext *context);
 	void skip();
 	std::string asString(DynamicContext *context, int indent) const;
@@ -65,8 +65,8 @@ private:
 	bool todo_;
 	Result parent_;
 	const ASTNode *pred_;
-	unsigned int contextPos_;
-	unsigned int contextSize_;
+	size_t contextPos_;
+	size_t contextSize_;
 	Item::Ptr first_;
 	Item::Ptr second_;
 };
@@ -74,7 +74,7 @@ private:
 class NonNumericPredicateFilterResult : public ResultImpl
 {
 public:
-	NonNumericPredicateFilterResult(const Result &parent, const ASTNode *pred, unsigned int contextSize);
+	NonNumericPredicateFilterResult(const Result &parent, const ASTNode *pred, size_t contextSize);
 	Item::Ptr next(DynamicContext *context);
 	void skip();
 	std::string asString(DynamicContext *context, int indent) const;
@@ -82,14 +82,14 @@ private:
 	bool todo_;
 	Result parent_;
 	const ASTNode *pred_;
-	unsigned int contextPos_;
-	unsigned int contextSize_;
+	size_t contextPos_;
+	size_t contextSize_;
 };
 
 class NumericPredicateFilterResult : public ResultImpl
 {
 public:
-	NumericPredicateFilterResult(const Result &parent, const ASTNode *pred, unsigned int contextSize);
+	NumericPredicateFilterResult(const Result &parent, const ASTNode *pred, size_t contextSize);
 	Item::Ptr next(DynamicContext *context);
 	void skip();
 	std::string asString(DynamicContext *context, int indent) const;
@@ -97,7 +97,7 @@ private:
 	bool todo_;
 	Result parent_;
 	const ASTNode *pred_;
-	unsigned int contextSize_;
+	size_t contextSize_;
 };
 
 #endif
